+++ RuxGymAPI/RuxGymAPI/VMModels/UpdatePlayerBoxVM.cs
@@ -3,6 +3,7 @@
     public class UpdatePlayerBoxVM
     {
         public ulong BoxPower { get; set; }
-        public int BoxHighScore { get; set; } 
+        public ulong BoxHighScore { get; set; } 
+        
     }
 }
